feat(pictl): waves 3-10 — full autonomics pipeline completion
10-wave van der Aalst agent swarm delivering comprehensive process
mining WASM library hardening:

- Wave 3: Zero clippy warnings, cargo fmt across 93 source files
- Wave 4: Test coverage expansion with parity + integration tests
- Wave 5: Criterion benchmarks for all algorithm tiers
- Wave 6: OCEL feature hardening (flatten, conformance, performance)
- Wave 7: POWL full pipeline (discovery, simplification, playout)
- Wave 8: Streaming algorithm hardening (SIMD DFG, alpha, heuristic)
- Wave 9: ML prediction pipeline hardening (drift, outcome, resource)
- Wave 10: Final audit — compilation clean, 69 tests passing

Key fixes:
- extensive_playout: can_emit flag for proper sequence trace emission
- montecarlo: test fixture matching num_cases parameter
- alignment_fitness: preset/postset indexed by transition not place
- A* alignments: floating-point f_cost comparison without truncation

175 files changed, 21,452 insertions, 1,422 deletions

+++ wasm4pm/src/powl_parser.rs
@@ -181,7 +181,7 @@ fn parse_decision_graph(s: &str, arena: &mut PowlArena) -> Result<u32, String> {
                     .position(|(t, _)| node_label_matches(t, tgt_str))
                     .ok_or_else(|| format!("edge target '{}' not found in nodes", tgt_str))?;
 
-                order.add_edge(src_local as usize, tgt_local as usize);
+                order.add_edge(src_local, tgt_local);
             }
         }
     }
@@ -211,7 +211,7 @@ fn parse_node_list(s: &str, token_to_local: &[(String, u32)]) -> Result<Vec<usiz
             .iter()
             .position(|(t, _)| node_label_matches(t, tok))
             .ok_or_else(|| format!("node '{}' not found in token list", tok))?;
-        indices.push(idx as usize);
+        indices.push(idx);
     }
     Ok(indices)
 }
